docs: README.md 덮어쓰기 refactor: app폴더 안의 file들의 page.tsx에 각각의 page구현, 		components 폴더 안의 common폴더에 Mainnav.tsx, Footer.tsx 위치, 		landing 폴더 안에 랜딩페이지 안에 들어갈 요소들 위치, 		upload 폴더 안에 uploadPage 안에 들어갈 요소들 위치 예정, 		sign 폴더 안에 logInPage, signUpPage 안에 들어갈 요소들 위치 예정
