--- src/app/main/page.tsx
@@ -1,18 +1,0 @@
-import MainPage from "@/components/landingPage/MainPage";
-
-function Page() {
-  return (
-    <>
-      <MainPage />
-      {/*<Header/>*/}
-      {/*<MainDescription/>*/}
-      {/*<UseCase/>*/}
-      {/*<GetItATry/>*/}
-      {/*<Footer/>*/}
-
-    </>
-  );
-}
-export default Page;
-
-// ㅁㄹㄴㅇㅇㅁㄴㄹㄴㅁㅇㄹ
