+++ src/components/common/Mainnav.tsx
@@ -15,6 +15,3 @@ function Mainnav() {
 }
 
 export default Mainnav;
-
-//asfsdaf
-//asdf
